+++ CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - [Utility] Non-resource files may be loaded using utility functions. [#235]
 
+- [Schemas] Test that multiple Rulesets can be added to a Schema. [#254]
+
 - [Validation] `full_validation()` now checks whether a Dataset is IATI XML. [#239]
 - [Validation] Test that SSOT organisation test files are valid IATI XML. [#242]